Split receiver header up
